backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
bcrypt==4.1.2

# Redis
redis==5.0.1
hiredis==2.3.2

# Machine Learning Core
torch==2.1.2
torchvision==0.16.2
onnxruntime==1.16.3
numpy==1.24.3
opencv-python==4.9.0.80
scikit-learn==1.4.0
scipy==1.11.4

# Face Recognition Models
insightface==0.7.3
onnx==1.15.0
facenet-pytorch==2.5.3

# Vector Search
faiss-cpu==1.7.4

# HTTP Client
httpx==0.26.0
aiofiles==23.2.1

# Utilities
python-json-logger==2.0.7

#Face recognition libraries

opencv-python-headless
deepface
tf-keras
pillow